Update souffle-header files
