+++ cbits/souffle/datastructure/BTreeDelete.h
@@ -904,8 +904,10 @@ public:
     /**
      * The iterator type to be utilized for scanning through btree instances.
      */
-    class iterator : public std::iterator<std::bidirectional_iterator_tag, Key> {
-    public:
+    class iterator {
+        friend class souffle::detail::btree_delete<Key, Comparator, Allocator, blockSize, SearchStrategy,
+                true, WeakComparator, Updater>;
+
         // a pointer to the node currently referred to
         // node const* cur;
         node* cur;
@@ -913,6 +915,7 @@ public:
         // the index of the element currently addressed within the referenced node
         field_index_type pos = 0;
 
+    public:
         using iterator_category = std::forward_iterator_tag;
         using value_type = Key;
         using difference_type = ptrdiff_t;
@@ -1578,7 +1581,7 @@ public:
         } else {
             auto lower_iter = internal_lower_bound(k);
             if (lower_iter != end() && equal(*lower_iter, k)) {
-                return distance(lower_iter, internal_upper_bound(k));
+                return std::distance(lower_iter, internal_upper_bound(k));
             } else {
                 return 0;
             }
@@ -1605,7 +1608,7 @@ public:
         } else {
             iterator lower_iter = internal_lower_bound(k);
             if (lower_iter != end() && equal(*lower_iter, k)) {
-                size_type count = distance(lower_iter, internal_upper_bound(k));
+                size_type count = std::distance(lower_iter, internal_upper_bound(k));
                 for (size_type i = 0; i < count; i++) {
                     erase(lower_iter);
                 }
