+++ cbits/souffle/io/ReadStreamCSV.h
@@ -347,8 +347,8 @@ protected:
      */
     static std::string getFileName(const std::map<std::string, std::string>& rwOperation) {
         auto name = getOr(rwOperation, "filename", rwOperation.at("name") + ".facts");
-        if (name.front() != '/') {
-            name = getOr(rwOperation, "fact-dir", ".") + "/" + name;
+        if (!isAbsolute(name)) {
+            name = getOr(rwOperation, "fact-dir", ".") + pathSeparator + name;
         }
         return name;
     }
