+++ cbits/souffle/io/ReadStreamSQLite.h
@@ -127,6 +127,7 @@ protected:
     }
 
     void openDB() {
+        sqlite3_config(SQLITE_CONFIG_URI, 1);
         if (sqlite3_open(dbFilename.c_str(), &db) != SQLITE_OK) {
             throwError("SQLite error in sqlite3_open: ");
         }
@@ -170,6 +171,10 @@ protected:
         // convert dbname to filename
         auto name = getOr(rwOperation, "dbname", rwOperation.at("name") + ".sqlite");
         name = getOr(rwOperation, "filename", name);
+
+        if (name.rfind("file:", 0) == 0 || name.rfind(":memory:", 0) == 0) {
+            return name;
+        }
 
         if (name.front() != '/') {
             name = getOr(rwOperation, "fact-dir", ".") + "/" + name;
