+++ cbits/souffle/io/WriteStreamSQLite.h
@@ -65,9 +65,11 @@ protected:
             }
 
 #if RAM_DOMAIN_SIZE == 64
-            if (sqlite3_bind_int64(insertStatement, i + 1, value) != SQLITE_OK) {
+            if (sqlite3_bind_int64(insertStatement, static_cast<int>(i + 1),
+                        static_cast<sqlite3_int64>(value)) != SQLITE_OK) {
 #else
-            if (sqlite3_bind_int(insertStatement, i + 1, value) != SQLITE_OK) {
+            if (sqlite3_bind_int(insertStatement, static_cast<int>(i + 1), static_cast<int>(value)) !=
+                    SQLITE_OK) {
 #endif
                 throwError("SQLite error in sqlite3_bind_text: ");
             }
@@ -102,7 +104,7 @@ private:
         throw std::invalid_argument(error.str());
     }
 
-    uint64_t getSymbolTableIDFromDB(int index) {
+    uint64_t getSymbolTableIDFromDB(std::size_t index) {
         if (sqlite3_bind_text(symbolSelectStatement, 1, symbolTable.decode(index).c_str(), -1,
                     SQLITE_TRANSIENT) != SQLITE_OK) {
             throwError("SQLite error in sqlite3_bind_text: ");
@@ -115,7 +117,7 @@ private:
         sqlite3_reset(symbolSelectStatement);
         return rowid;
     }
-    uint64_t getSymbolTableID(int index) {
+    uint64_t getSymbolTableID(std::size_t index) {
         if (dbSymbolTable.count(index) != 0) {
             return dbSymbolTable[index];
         }
@@ -140,8 +142,9 @@ private:
     }
 
     void openDB() {
+        sqlite3_config(SQLITE_CONFIG_URI, 1);
         if (sqlite3_open(dbFilename.c_str(), &db) != SQLITE_OK) {
-            throwError("SQLite error in sqlite3_open");
+            throwError("SQLite error in sqlite3_open: ");
         }
         sqlite3_extended_result_codes(db, 1);
         executeSQL("PRAGMA synchronous = OFF", db);
@@ -269,6 +272,10 @@ private:
         // convert dbname to filename
         auto name = getOr(rwOperation, "dbname", rwOperation.at("name") + ".sqlite");
         name = getOr(rwOperation, "filename", name);
+
+        if (name.rfind("file:", 0) == 0 || name.rfind(":memory:", 0) == 0) {
+            return name;
+        }
 
         if (name.front() != '/') {
             name = getOr(rwOperation, "output-dir", ".") + "/" + name;
