+++ package.yaml
@@ -59,9 +59,6 @@ install-includes:
   - souffle/CompiledSouffle.h
   - souffle/RamTypes.h
   - souffle/RecordTable.h
-  - souffle/datastructure/ConcurrentFlyweight.h
-  - souffle/datastructure/ConcurrentInsertOnlyHashMap.h
-  - souffle/utility/ParallelUtil.h
   - souffle/utility/span.h
   - souffle/SignalHandler.h
   - souffle/SouffleInterface.h
@@ -71,18 +68,23 @@ install-includes:
   - souffle/utility/Iteration.h
   - souffle/utility/Types.h
   - souffle/utility/tinyformat.h
-  - souffle/utility/StreamUtil.h
-  - souffle/utility/ContainerUtil.h
-  - souffle/utility/DynamicCasting.h
   - souffle/datastructure/BTreeDelete.h
   - souffle/datastructure/BTreeUtil.h
   - souffle/utility/CacheUtil.h
+  - souffle/utility/ContainerUtil.h
+  - souffle/utility/DynamicCasting.h
+  - souffle/utility/ParallelUtil.h
   - souffle/datastructure/Brie.h
+  - souffle/utility/StreamUtil.h
   - souffle/datastructure/EquivalenceRelation.h
   - souffle/datastructure/LambdaBTree.h
   - souffle/datastructure/BTree.h
   - souffle/datastructure/PiggyList.h
   - souffle/datastructure/UnionFind.h
+  - souffle/datastructure/RecordTableImpl.h
+  - souffle/datastructure/ConcurrentFlyweight.h
+  - souffle/datastructure/ConcurrentInsertOnlyHashMap.h
+  - souffle/datastructure/SymbolTableImpl.h
   - souffle/datastructure/Table.h
   - souffle/io/IOSystem.h
   - souffle/io/ReadStream.h
